Move shuffle functions to new file and put some symbols in the pdns namespace.
